@@ -6,31 +6,33 @@
         "ext-ctype": "*",
         "ext-iconv": "*",
         "ext-json": "*",
+        "exercise/htmlpurifier-bundle": "^3.0",
         "sensio/framework-extra-bundle": "^5.5",
-        "symfony/asset": "5.0.*",
-        "symfony/console": "5.0.*",
-        "symfony/dotenv": "5.0.*",
-        "symfony/expression-language": "5.0.*",
+        "symfony/asset": "5.1.*",
+        "symfony/console": "5.1.*",
+        "symfony/dotenv": "5.1.*",
+        "symfony/expression-language": "5.1.*",
         "symfony/flex": "^1.3.1",
-        "symfony/form": "5.0.*",
-        "symfony/framework-bundle": "5.0.*",
-        "symfony/http-client": "5.0.*",
-        "symfony/intl": "5.0.*",
-        "symfony/mailer": "5.0.*",
+        "symfony/form": "5.1.*",
+        "symfony/framework-bundle": "5.1.*",
+        "symfony/http-client": "5.1.*",
+        "symfony/intl": "5.1.*",
+        "symfony/mailer": "5.1.*",
         "symfony/monolog-bundle": "^3.1",
-        "symfony/notifier": "5.0.*",
+        "symfony/notifier": "5.1.*",
         "symfony/orm-pack": "*",
-        "symfony/process": "5.0.*",
-        "symfony/security-bundle": "5.0.*",
+        "symfony/process": "5.1.*",
+        "symfony/security-bundle": "5.1.*",
         "symfony/serializer-pack": "^1.0",
-        "symfony/string": "5.0.*",
-        "symfony/translation": "5.0.*",
+        "symfony/string": "5.1.*",
+        "symfony/translation": "5.1.*",
         "symfony/twig-pack": "*",
-        "symfony/validator": "5.0.*",
-        "symfony/web-link": "5.0.*",
-        "symfony/yaml": "5.0.*"
+        "symfony/validator": "5.1.*",
+        "symfony/web-link": "5.1.*",
+        "symfony/yaml": "5.1.*"
     },
     "require-dev": {
+        "roave/security-advisories": "dev-master",
         "doctrine/doctrine-fixtures-bundle": "^3.3",
         "symfony/debug-pack": "*",
         "symfony/maker-bundle": "^1.0",
@@ -80,7 +82,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.0.*"
+            "require": "5.1.*"
         }
     }
 }
