@@ -1,86 +1,77 @@
 <template>
-    <div class="form-group">
-        <label class="typo__label" for="ajax">Kontakt</label>
-        <multiselect
-                v-model="model"
-                :value="model"
-                id="ajax"
-                label="name"
-                track-by="id"
-                placeholder="Type to search"
-                open-direction="bottom"
-                :options="preselectedOptions"
-                :multiple="false"
-                :searchable="true"
-                :loading="isLoading"
-                :internal-search="false"
-                :close-on-select="true"
-                :options-limit="300"
-                :limit="3"
-                :limit-text="limitText"
-                :max-height="600"
-                :show-no-results="false"
-                :hide-selected="true"
-                :preserveSearch="true"
-                @search-change="defaultSearch"
-        >
+  <div class="form-group">
+    <label class="typo__label" for="ajax">Kontakt</label>
+    <multiselect
+        v-model="value"
+        id="ajax"
+        label="name"
+        track-by="id"
+        placeholder="Type to search"
+        open-direction="bottom"
+        :options="preselectedOptions"
+        :multiple="false"
+        :searchable="true"
+        :loading="isLoading"
+        :internal-search="false"
+        :close-on-select="true"
+        :options-limit="300"
+        :limit="3"
+        :limit-text="limitText"
+        :max-height="600"
+        :show-no-results="false"
+        :hide-selected="true"
+        :preserveSearch="true"
+        @search-change="defaultSearch"
+        @input="$emit('input', value)"
+    >
 
             <span slot="noResult">
                 Oops! No elements found. Consider changing the search query.
             </span>
-        </multiselect>
-
-        <input type="hidden" :id="id" :name="name" :value="model.id">
-    </div>
+    </multiselect>
+  </div>
 </template>
 <script>
-    import axios from 'axios';
-    import Multiselect from 'vue-multiselect';
+import axios from 'axios';
+import qs from 'qs';
+import Multiselect from 'vue-multiselect';
+import 'vue-multiselect/dist/vue-multiselect.min.css';
 
-    export default {
-        components: {
-            Multiselect,
-        },
-        props: {
-            id: { type: String, required: true },
-            name: { type: String, default: '' },
-            value: { type: Object, default: () => {} },
-            url: { type: String, required: true },
-        },
-        data () {
-            return {
-                isLoading: false,
-                preselectedOptions: [],
-            }
-        },
-        computed: {
-            model: {
-                get() {
-                    return this.value;
-                },
-                set(val) {
-                    this.$emit('input', val)
-                }
-            },
-        },
-        methods: {
-            limitText (count) {
-                return `and ${count} other countries`
-            },
-            defaultSearch(query) {
-                this.isLoading = true;
-
-                axios({
-                    method: 'get',
-                    url: this.url + query,
-                }).then(response => {
-                    this.preselectedOptions = response.data;
-                    this.isLoading = false;
-                }).catch(error => {
-                    this.preselectedOptions = [];
-                    this.isLoading = false
-                });
-            }
-        }
+export default {
+  components: {
+    Multiselect,
+  },
+  props: {
+    name: {type: String, default: ''},
+    url: {type: String, required: true},
+  },
+  data() {
+    return {
+      value: '',
+      isLoading: false,
+      preselectedOptions: [],
     }
+  },
+  methods: {
+    limitText(count) {
+      return `and ${count} other countries`
+    },
+    defaultSearch(query) {
+      this.isLoading = true;
+
+      axios({
+        method: 'post',
+        url: this.url,
+        headers: {'content-type': 'application/x-www-form-urlencoded'},
+        data: qs.stringify({name: query,}),
+      }).then(response => {
+        this.preselectedOptions = response.data;
+        this.isLoading = false;
+      }).catch(error => {
+        this.preselectedOptions = [];
+        this.isLoading = false
+      });
+    }
+  }
+}
 </script>
