@@ -4,7 +4,7 @@ namespace App\Repository\Transaction;
 
 use App\Entity\Transaction\Row;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ManagerRegistry;
 
 /**
  * @method Row|null find($id, $lockMode = null, $lockVersion = null)
